+++ src/ElevateOTT.Application/Features/Content/Videos/Queries/GetVideos/VideoItem.cs
@@ -14,10 +14,10 @@ public class VideoItem : BaseAssetDto
     public Guid? AuthorId { get; set; }
 
     public bool HasOneTimePurchasePrice { get; set; }
-    public decimal OneTimePurchasePrice { get; set; }
+    public double OneTimePurchasePrice { get; set; }
     public bool HasRentalPrice { get; set; }
     public RentalDuration RentalDuration { get; set; }
-    public decimal RentalPrice { get; set; }
+    public double RentalPrice { get; set; }
 
     public AuthorDto? Author { get; set; }
     public List<AssetImageDto>? VideoImages { get; set; }
